de-i18n analytics and add r.js config
config/build.js tells r.js how to build the optimized javascript for the
plugin. the mess of ugly paths is because of all the javascript from
canvas-lms that's used by the plugin (really, just by common and
handlebars_helpers).

I18n doesn't work yet in plugins. we had included the I18n calls anyways
since we believed they were no-ops rather than a hangup (and in
development mode, that's true). but when r.js tries to optimize the
javascript, it wants translation files -- which we don't have, until
I18n in the plugin works.

Change-Id: Ib0ca989ac97c75d9b791c54f38bc07f91fecb982
Reviewed-on: https://gerrit.instructure.com/9292
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>