fix hudson_setup.sh
hudson only clears out plugins that it knows about between builds. since
it thinks this plugin is canvalytics, it doesn't remove the
vendor/plugins/analytics directory. then when it tries to move the
canvalytics checkout to analytics, it moves it inside instead. then, on
the third and subsequent runs, it fails because
vendor/plugins/analytics/canvalytics is already there.

before trying to move canvalytics -> analytics, remove any old copy of
analytics

Change-Id: I7fb2200251a9b39e1109f13589992c0910ef4ab4
Reviewed-on: https://gerrit.instructure.com/9299
Reviewed-by: Cameron Matheson <cameron@instructure.com>
Tested-by: Hudson <hudson@instructure.com>