add more non-local files to config/build.js
these are new files (or groups of files) added to canvas-lms which are
now pulled in to analytics via handlebars_helpers (which is required by
all jst files). this type of fix is *NOT IDEAL*, but we need to get the
deploy out. we really need to figure out a better way to manage this.

Change-Id: I71cee9ebab3b6ca10d27cce61d1752d3e934916c
Reviewed-on: https://gerrit.instructure.com/9613
Tested-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>