overhaul date handling in graphs, fixes #8357
refactor date_axis from a function applied to relevant graphs to a base
class those graphs can inherit from. as a prerequisite, refactor padding
parameters from the individual graphs into the base graph class.

eliminate binning of days as integers, especially the round tripping
that caused the bug. in the process, move data organization from the
graph classes to the data classes' populate methods.

make sure if the start/end dates fall in the middle of a day that we
choose the start/end midnights to be inclusive.

test-plan:
  - load analytics page
  - hover a page view bar to display tooltip
  - date in tooltip should match date from x-axis
  - repeat for other date-aligned graphs on the student-in-course and
    course pages

Change-Id: I9dc42fa6c0a01c5ddb02845f3c14569a35c47c90
Reviewed-on: https://gerrit.instructure.com/10518
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>