analytics rollups
CachedGradeDistribution counts the number of students in a course with
each grade (rounded to nearest integer, ignoring grades outside 0..100).
It is recalculated for a course any time a student is added to the
course, an enrollment state changes (e.g. invited -> active, active ->
deleted affects the counts), or a student's grade is recalculated via
the GradeCalculator.

PageViewsRollup counts the number of page views and the number of
participations in a course on a given day for a given controller
category. Once migrated, the rollup is kept up to date by incrementing
the appropriate bin every time a page view is created.

test-plan:
 - run migrations (WARNING: the page view rollup might take a while
   depending on how many page views you have)
 - spot check some rollup rows vs. real data
 - add an active student to a course; the "s0" column of the course's
   cached_grade_distribution should increment
 - change the student's score by e.g. grading a submission; the cached
   distribution should update appropriately.
 - view some pages; verify the appropriate rollup bins get incremented
 - "participate" (e.g. reply in a discussion); verify the appropriate
   rollup bin gets its participations incremented

Change-Id: I86c0c566f82063edfaf5ca26a13f4a09c8916708
Reviewed-on: https://gerrit.instructure.com/10939
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>