truncate student list in course analytics
if a course has hundreds (or thousands, for a MOOC) of students, trying
to load summaries for every student in the course will take way to long.

eventually, we'll paginate this and use pageless and yadda yadda, but
for now, just truncate the student list to the first 50 students, and
add some text telling them how to get to individual student analytics
for the other students.

fixes #9031

test-plan:
  - view course analytics for a course with fewer than 50 students.
  - should behave just as before
  - view course analytics for a course with more than 50 students.
  - should see the first 50 students (alphabetically), but no more.
  - after the student list there should be a notice telling them it was
    truncated.

Change-Id: Ife0b780f18c8c9a78057909a6703cd167aa4d8d9
Reviewed-on: https://gerrit.instructure.com/11602
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>