tweak to go along with outcomes refactor
the recent outcomes refactor included temporary shims to let old code in
plugins work, but now that the refactor is on master, we can fix this
plugin to move off the shims. once this and like commits in other
plugins have been submitted, the shims included with the refactor can be
removed.

test-plan:
 - specs pass

Change-Id: I3e5f78d190197d43c38f60cccc95cf468e20708e
Reviewed-on: https://gerrit.instructure.com/12796
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>