fix analytics caching by visibility; fixes #10726
when caching analytics data, take visibility into consideration so that
a TAs restricted view isn't cached for later teachers/admins (annoying
and wrong) and a teacher's unrestricted view isn't cached for later TAs
(worse, leaks data).

no account-level statistics are dependent on viewer, so no change there.
almost all course-level statistics are dependent on which enrollments
the viewer can see, so add that to the base cache key for courses.
assignment data at both the course- and student-levels are dependent on
whether the viewer can see student details, so include that in the cache
key for that specific item.

test-plan:

 * enable redis caching
 * create a course with a teacher, two sections, one student per
   section, and a TA who's limited to the first section
   - clear the redis cache
   - as the teacher, view course-level analytics for the course
   - values should reflect both students
   - as the TA, view course-level analytics for the course
   - values should reflect only the student from the TA's section
 * create a course with a teacher and a student
   - create an assignment in the course and mute it
   - as the student, make a submission to the assignment
   - as the teacher, grade the submission (assignment still muted)
   - clear the redis cache
   - as the teacher, view student-level analytics for the student
   - the assignment should show normally
   - as the student, view student-level analytics for the student
   - the assignment should show as muted

Change-Id: I41b8722667605ae56b2e3b1f7157c678aa71971c
Reviewed-on: https://gerrit.instructure.com/13821
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>