update analytics for cassandra page views
refs #10756

this depends on the corresponding canvas-lms commit, which maintains
compatibility with AR page view store as well. so this commit supports
both back-ends, though someday we may drop the AR support.

test plan: things should work as before. make sure to have the
canvas-lms change as well. test with page views configured for both
postgresql and for cassandra.

Change-Id: I0d4e9eb0f1d0d609bac9529065d2ceb7004682e1
Reviewed-on: https://gerrit.instructure.com/14269
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>