Fix grade distribution cache recalculation
This was throwing a "missing attribute: user_id"
every time the recalculation fired on my machine.

the recalculate! method was running 
a find statement that specified which columns 
were selected, and user_id was not among them,
so the association preload (coming from the 
all_student_enrollments association [ which
used :include=>:user] on the
enrollment model)

Confirmed with Jacob that I wasn't missing 
anything dependency-wise, and agreed on the fix. 
Now this just executes some sql and gets back raw 
rows of data because we don't really need an 
enrollment object here and this avoids 
the preload entirely.

Change-Id: I5800bbb5849283443d37cb9026d042e1ccb62742
Reviewed-on: https://gerrit.instructure.com/14890
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>