refs #9171 account level analytics button
this used to check for the
right column buttons using the
existential operator in coffeescript.
In jquery, if an element is not found on
the page than it is just an empty array,
which will still pass an existance check.

I've changed the coffeescript to check
the length of the array produced by
the jquery selector instead,
and to take the remedial action if
that length is 0.

also, this didn't seem big
enough to justify an entire
selenium test to verify, so I pushed
the faulty logic into it's own mini
module and wrapped some qunit around
it.

Change-Id: Ieaeb682d9fb3308638c4b0e6af80142bd19c313a
Reviewed-on: https://gerrit.instructure.com/14988
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>