refs #9171 fix analytics permissions
the permissions list being passed to
authorized_action was not being
wrapped in array brackets and this
was only allowing one of the
specified permissions to get through to
the actual permissions check.

Adding the array brackets fixed it,
and I added a spec file for
the controller since it didn't have
one yet.  My specs for this
are a bit verbose as I was more
interested in testing the specifics
of the permissions passing than
the overall rendering.

Change-Id: I39af7d471acb71ecc5726c0d1114eb5afb5cd035
Reviewed-on: https://gerrit.instructure.com/15062
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>