don't index paginated students by all student_ids
this had used a local variable students_ids = students.map(&:id), but
that assignment was removed, and it accidentally kept using the
student_ids method. this wasn't a problem when the local variable
students matches the method students. but when there were more than 50
students in the course, students != self.students, and self.student_ids
included ids not in summaries.keys, causing an error.

fixes #11961

test-plan:
 - create a course with more than 50 active students
 - load the course analytics page
 - should see the first 50 student summaries

Change-Id: I58310befb95b60df824249e41553967793ba7096
Reviewed-on: https://gerrit.instructure.com/15618
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Brian Palmer <brianp@instructure.com>