don't use Times as keys of json-destined Hash
when a ruby Hash is converted to json, it uses to_s on the keys instead
of to_json, since all json hash keys are strings. if the keys of the
Hash are Time objects, this produces human formatted keys instead of the
required iso8601 (which to_json would have produced if it were a value
instead of a key).

fixes CNVS-3408

test-plan:
  - be using cassandra page views
  - have a student in a course
  - produce some pages views for that student/course
  - fetch the activity data for that student-in-course from the api
    (/api/v1/courses/X/analytics/users/Y/activity)
  - the keys of the page_views object should be iso8601 formatted

Change-Id: I7317d7af374b585d376c9a7c9d6bd98b427ac55a
Reviewed-on: https://gerrit.instructure.com/17277
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>
Tested-by: Jacob Fugal <jacob@instructure.com>