keep scope for submissions down to section level
REFS CNVS-5720

this previously did not allow rollups to just load
submissions for a certains section, which would produce both incorrect
data and (possibly) drag down the speed of the job run
(though the incorrect data is more important in this case).

spec has been added to ensure it's only pulling data from the
specified section.

TEST PLAN:

This one is hard to test unless you can get down at the data.

1) kill all assignment rollups (AssignmentRollup.delete_all)
2) Run the assignment rollup job (AssignmentsRoller.rollup_all)
3) check the rollups (AssignmentRollup.last and friends), the ones that have a "course_section_id"
should only have data that reflects the submissions for that section, not for the entire course.
(for example, if there are 2 sections, and each has one participant, and one scored a 10
on a given assignment, and one scored a 1, then the minimum score for the rollup for the higher
scoring section should be 10, and before it would have been 1 because it would have been pulling
all submission data regardless of course section)

Change-Id: Id331dd3be1161f8ca191a2212671ad9a38c7c63d
Reviewed-on: https://gerrit.instructure.com/20433
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
QA-Review: Cam Theriault <cam@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>