remove assignment rollup db table and optimize rollup generation
refs CNVS-5805

We gave up on storing the rollup because it implicitly relies on the current
time at generation, and can go stale at any point based on numerous due
dates. Instead, we've heavily optimized the rollup generation, and we're
going to generate it on-demand and cache it.

This commit generates it synchronously on-demand in the api request. A
subsequent commit will switch to async generation if an async param is
passed to the api.

test plan: the assignments graph on the course analytics page should
continue to display correct data, including for section-limited teachers
and students.

Change-Id: I096c8a1a6f431bee03f6ad1974de741693d31368
Reviewed-on: https://gerrit.instructure.com/20841
Reviewed-by: Brian Palmer <brianp@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>
Tested-by: Jacob Fugal <jacob@instructure.com>