analytics showing submissions were submitted if teacher give a 0
refs: CNVS-1601

Since submissions were not created for students who did not submit an
assignment the analytics were showing them as missing even if they were not due
yet. Regardless what the score was.  The fix changed the DueDateCacher to create
submissions for students with overridden due dates.  This will cache the due date
so analyitics can correctly calculate the graph data.  The analytics will search
for a submission,  if it finds one it will use the cached due date for
calculations,  if not it uses the assignments date due assuming there are not any
dates that are overridden.

Also reference g/23800.

Test Case:
  Setup: Create a Course with an assignment and an enrollment.
  - Without the student submitting the assignment. Grade the assignment for the
    student.
  - Open analytics for the course and the Assignments graph should not show any
    submissions for that assignment.
  - Create a due date override.
  - A submission should be created to cache the due date and the graph should
    not show any submissions for that assignment.

Change-Id: I17ba5f42d41da3970435f06fa2889598390b696d
Reviewed-on: https://gerrit.instructure.com/26131
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>