changes to user participations api response
closes CNVS-25068

See the API doc changes, we are now returning participation counts in
the same format as the page view counts, rather than including details
on each participation.

Note that this is a breaking API change. We try hard not to do this, but
we won't be storing this information anymore due to the high cost and
low utility.

test plan:
enable analytics and go to an individual student-in-course analytics
page. the top graph of the student's page views and participations
should still display as before. you can also hit the api directly and
test that the response now matches the changed api documentation.

Change-Id: Ibc24b2a43c34e5cacbe308a4d7f658c1073eb43b
Reviewed-on: https://gerrit.instructure.com/66830
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>