consistently convert timestamps into dates
fixes CNVS-27694

the query would cast to UTC (though still not a date, but postgres
handled that conversion itself) automatically, but the assignment to
date on a new record would not, grabbing the date portion according to
the embedded timezone. so if, e.g., the incoming timestamp was for 9pm
MDT, the query would look for the UTC day (3am UTC), but after failing
would try and build a bin for the MDT day, which bin probably already
existed.

Change-Id: I216c0b9e4a4d197db8fdcc8e5c0473f1593a1f23
test-plan: specs
Reviewed-on: https://gerrit.instructure.com/83161
Reviewed-by: Rob Orton <rob@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>