Move toggle label into aria-label
Moving the toggle/checkbox's label from a screenreader-only span to the
checkbox's aria-label does the trick.

NOTE: the change is in the analytics repo. to test, checkout the gerrit
into gems/plugins/analytics!

fixes ADMIN-1313

test plan:
  - have a student in a course with some activity
  - go to /accounts/1/analytics
  > expect screenreaders to read the page, including the view toggle,
    correctly
  > expect JAWS+ie to include the toggle's label in the menu list of
    form fields (keybad-insert + f5).

  - repeat at /courses/:course_id/analytics/users/:user_id (go to the
    course -> people, click on a student, in the student tray that opens
    click on Analytics)
  > expect it to work here too

  - repeat for department analytics, but I don't know where that is
    (though all the changes are exactly the same, so you could assume
    it works here too if it worked in the other 2 places)

In firefox, jaws doesn't see _any_ form fields on this page, or
the edit assignment page called out in the ticket as an example
of where to see the menu work properly. This is an issue with
ff + jaws, and not the ticket.

Change-Id: I4919a4df1d2c97f6e593b1eebc6362e26cfb7b56
Reviewed-on: https://gerrit.instructure.com/161159
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>