Fix merging of participation and page views
the binning of page views and participation from the api into dates is
fixed.

closes ADMIN-1299

note: there was a change in this file from binning by day -> hour to fix
a bug that looks like was a timezone issue. See CNVS-16464 and
https://gerrit.instructure.com/#/c/44111/.  I believe this sort of
addressed the issue at the time, but was not correct.

test plan:
  - have a course with a student with page views and participation data
    on the same day (you could run @jeremy's fake_analytics.rb script to
    generate randomish data and tweak the results to get dupes)
  - go to /courses/:couse_id/analytics/users/:user_id
  - click the toggle to switch to table view
  > expect the activity by date table to have combined the views and
    participation into the same date. There should be no duplicate
    dates in the table.

Change-Id: Iff76a43c0a7c0a1d3dd5082a6bbde18407916f32
Reviewed-on: https://gerrit.instructure.com/161479
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Carl Kibler <ckibler@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>